common: bbu: only add available handlers
A board may have multiple handlers registered:

  registered update handlers:
    SD          -> /dev/mmc0.barebox
  * eMMC        -> /dev/mmc1

However when using the usbgadget with the -b command line argument,
fastboot will stat all devicefiles listed in the handlers and will fail
if i.e. the SD card is not available:

  usbgadget -A /dev/mmc1(root) -b
  udc0: registering UDC driver [g_multi]
  multi_bind: creating Fastboot function
  ERROR: g_multi udc0: failed to start g_multi: -2
  usbgadget: No such file or directory

To fix this, check the availability of handlers before adding them to
the list and skip those that are not available with an info level message.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>
Link: https://lore.barebox.org/20210618120557.2192098-1-r.czerwinski@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>