of: of_copy_node(): Copy phandles as well
of_copy_node() has to copy the phandles as well, because otherwise
the phandles in the copied tree become invalid.

This is not entirely correct because when of_copy_node() is used to
copy one tree as a subtree of another tree then we might get duplicated
phandles which will result in a corrupted tree. This is also an issue
already without this patch, because the phandle of a node is not only
stored in the phandle member of a struct device_node, but also as a
separate phandle property in the tree itself. The phandle property is
copied before this patch, so we can equally well also copy the phandle
member of struct device_node.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>
Link: https://lore.barebox.org/20210624085223.14616-2-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>