commands: of_property: Extend command usage on external dtb
The scope of the of_property command is limited to the internal devicetree
only. There is the possibility to add fixtures to postpone the operation
until boot and after the kernel DTB is already loaded, but this does not
cover the case when the property value needs to be phandle adresses
defined in the external DTB, since assigned values are evaluated
immediatly.

This patch extends the of_property-command to use it on external DTB's as
well. In case fixup is not used, the changes will also affect the original
data.

Signed-off-by: Yunus Bas <y.bas@phytec.de>
Link: https://lore.barebox.org/20210625160555.3699702-1-y.bas@phytec.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>