habv4: habv4_get_status(): do not print HAB error event in case of known RNG self test failure
In commit

| 655e085e3a22 ("crypto: caam - Always do rng selftest")

the kconfig option CONFIG_CRYPTO_DEV_FSL_CAAM_RNG_SELF_TEST was
removed, which resulted in the pr_debug() never being executed. Remove
the IS_ENABLED() to restore the original behavior, i.e. to not print
the HAB error Event if it is the known RNG self-test failure.

Fixes: 655e085e3a22 ("crypto: caam - Always do rng selftest")
Reported-by: Roland Hieber <rhi@pengutronix.de>
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>
Reviewed-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>
Link: https://lore.barebox.org/20210716122110.4022409-1-mkl@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>