of: overlay: don't do fixups on empty global.of.overlay.dir
$global.of.overlay.dir is dynamically prepended with the root of the
device currently being booted. This allows use of relative paths.

Because $global.of.overlay.filepattern is * by default, this means that
all files in the root of the device will be considered overlays and
barebox will try to unflatten them leading to an error message on boot:

  ERROR: of_overlay: Failed to unflatten /mnt/mmc0.4//.root_is_rw:
	Invalid argument

Fix this by not doing any fixups at all on empty global.of.overlay.dir.

If for whatever reason, a user wants overlays at /, they can

  global.of.overlay.dir=.

Fixes: 124c64af0982 ("overlay: Add filters to choose which overlays to apply")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20210831161139.2988-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>